Stop truncating the player gait model's root node pointer
init_player_gait_object wrote

    player_gait_object_header.RootNode = (int)&player_gait_hdr;

RootNode is a ModelNode * and player_gait_hdr is a ModelNode, so the
cast was never needed; with 32-bit pointers it was also exact. On LP64
it truncates the address to its low half, and modelInitRwData was handed
0x39ad53c0 - the fault address, and the second argument in the register
dump - the moment the level tried to set up Bond's own model.

Drop the cast.

This is the stage the load now reaches. With the prop-def header and
zlib fixes in, the device run showed all 136 rooms inflating, the zlib
trap never firing, the converter and the game both walking 329 props,
173 objects created, and the intro cameras set up - the prop walk
stopped overrunning, so the last sane record is cmd=343 rather than
garbage at 536.

Swept for the same defect class with a grep for "= (int|s32|u32)&".
The only other live hits are src/game/bg.c:3804 and :3945, which assign
&D_800442FC[portalnum] to an s32 and then do "if (i);" - matching
artifacts whose value is never used. Left alone deliberately.

Not yet confirmed in the headset.

+++ src/game/initplayergaitobject.c
@@ -2,7 +2,12 @@
 #include "chrobjdata.h"
 
 void init_player_gait_object(void) {
-  player_gait_object_header.RootNode = (int)&player_gait_hdr;
+  /*
+   * RootNode is a ModelNode *, and player_gait_hdr is a ModelNode, so this
+   * needs no cast at all. The original wrote (int)&player_gait_hdr, which is
+   * exact with 32-bit pointers but on LP64 truncates the address to its low
+   * half: modelInitRwData was handed 0x39ad53c0 and faulted immediately.
+   */
+  player_gait_object_header.RootNode = &player_gait_hdr;
   return;
 }
-
